feat(story-round-two): Add story parts to response
This commit adds the story parts to the response for the /stories-round-2 endpoint. This allows for more accurate reads, votes, and comments to be displayed for each story.

The parts are sorted by create date to get the oldest part, and the reads, votes, and comments from the oldest part are used for the story response.

Authored-by: Lê Minh Trí <leminhtri11a1@gmail.com>